+++ pom.xml
@@ -7,7 +7,7 @@
   </parent>
 
   <groupId>com.puppetlabs.jenkins.plugins</groupId>
-  <artifactId>puppetgatlingjenkinsplugin</artifactId>
+  <artifactId>puppet-gatling-jenkins-plugin</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>hpi</packaging>
