--- src/main/java/com/puppetlabs/jenkins/plugins/puppetgatlingjenkinsplugin/Constants/Constant.java
@@ -1,18 +1,0 @@
-package com.puppetlabs.jenkins.plugins.puppetgatlingjenkinsplugin.Constants;
-
-/**
- * Constant
- * <br></br>
- * Location for important constant values related to the project
- * 
- * @author Brian Cain
- */
-
-public interface Constant {
-	String ICON_URL = "/plugin/customartifactbuilder/img/puppet.png";
-	String DISPLAY_NAME = "Puppet Gatling";
-	String URL_NAME = "puppet-gatling";
-
-	int MAX_BUILDS_TO_DISPLAY = 30;
-	int MAX_BUILDS_TO_DISPLAY_DASHBOARD = 15;
-}
