+++ src/main/resources/com/puppetlabs/jenkins/plugins/puppetgatling/PuppetGatlingProjectAction/floatingBox.jelly
