+++ src/main/resources/com/puppetlabs/jenkins/plugins/puppetgatling/PuppetGatlingProjectAction/index.jelly
