+++ src/main/resources/com/puppetlabs/jenkins/plugins/puppetgatling/PuppetGatlingPublisher/config.jelly
